face_cut/requirements.txt
boto3
requests
ydb
Pillow